moved config and related files to configuration directory
